ajout de la route get list by ID
