+++ .gitignore
@@ -8,3 +8,4 @@ easygroup.iws
 easygroup.ipr
 easygroup.iml
 logs
+.vscode
