+++ src/main/resources/application.properties
@@ -40,5 +40,5 @@ csrf.enabled=${CSRF_ENABLED:false}
 samesite.cookie.enabled=${SAMESITE_COOKIE_ENABLED:false}
 
 # JWT configuration
-jwt.secret=${JWT_SECRET:}
+jwt.secret=${JWT_SECRET:rKkTeex9w6EG9q4A4AGkpvgoz2DwP0wM6KhRWAK7fHw=}
 jwt.expiration=${JWT_EXPIRATION:86400000}
